REFACTOR domain format & signService folder change
