+++ src/main/java/snust/sbsp/crew/dto/req/EmailValidationReq.java
@@ -1,13 +1,14 @@
-package snust.sbsp.sign.dto.req;
+package snust.sbsp.crew.dto.req;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+
 @Getter
 @AllArgsConstructor
-public class SignInReq {
+public class EmailValidationReq {
 
   private String email;
 
-  private String password;
+  private String code;
 }
